revert change
Dumping the pct routes where no cycling is predicted under any senario. Tried inclusidn them but this had a terrible impact on performance and minimal affect on results. Typically around 40% of the lines are excluded but only 10 -15% of people, this also removes around 60% of the lines (40% people)  that cross regional boundaries reducing that problem too